+++ app/routes/app.$tenant/settings/api/keys/$id.tsx
@@ -5,6 +5,8 @@ import { i18nHelper } from "~/locale/i18n.utils";
 import UrlUtils from "~/utils/app/UrlUtils";
 import { useAppData } from "~/utils/data/useAppData";
 import { ApiKeyWithDetails, deleteApiKey, getApiKeyById, updateApiKey } from "~/utils/db/apiKeys.db.server";
+import { verifyUserHasPermission } from "~/utils/helpers/PermissionsHelper";
+import { getTenantUrl } from "~/utils/services/urlService";
 
 type LoaderData = {
   item: ApiKeyWithDetails;
@@ -26,6 +28,7 @@ type ActionData = {
 const badRequest = (data: ActionData) => json(data, { status: 400 });
 export const action: ActionFunction = async ({ request, params }) => {
   const { t } = await i18nHelper(request);
+  const tenantUrl = await getTenantUrl(params);
   const form = await request.formData();
   const action = form.get("action")?.toString() ?? "";
   const existing = await getApiKeyById(params.id ?? "");
@@ -33,6 +36,7 @@ export const action: ActionFunction = async ({ request, params }) => {
     return badRequest({ error: t("shared.notFound") });
   }
   if (action === "edit") {
+    await verifyUserHasPermission(request, "app.settings.apiKeys.update", tenantUrl.tenantId);
     const entities: { entityId: string; create: boolean; read: boolean; update: boolean; delete: boolean }[] = form
       .getAll("entities[]")
       .map((f: FormDataEntryValue) => {
@@ -49,10 +53,11 @@ export const action: ActionFunction = async ({ request, params }) => {
     );
     return redirect(UrlUtils.currentTenantUrl(params, "settings/api/keys"));
   } else if (action === "delete") {
+    await verifyUserHasPermission(request, "app.settings.apiKeys.delete", tenantUrl.tenantId);
     await deleteApiKey(params.id ?? "");
     return redirect(UrlUtils.currentTenantUrl(params, "settings/api/keys"));
   } else {
-    return badRequest(t("shared.invalidForm"));
+    return badRequest({ error: t("shared.invalidForm") });
   }
 };
 
@@ -64,7 +69,12 @@ export default function ApiEditKeyRoute() {
   return (
     <>
       <OpenModal className="sm:max-w-xl" onClose={() => navigate(UrlUtils.currentTenantUrl(params, "settings/api/keys"))}>
-        <ApiKeyForm entities={appData.entities} item={data.item} />
+        <ApiKeyForm
+          entities={appData.entities}
+          item={data.item}
+          canUpdate={appData.permissions.includes("app.settings.apiKeys.update")}
+          canDelete={appData.permissions.includes("app.settings.apiKeys.delete")}
+        />
       </OpenModal>
     </>
   );
